refactor(desktop): rm legacy helpers and mixins
- Move configManager to adapters, since it's a particular implementation
  of a config manager to handle ini files.
- Move gatewayMonitor to components, since it's specific for Qt
- Move utils to components, since it has functions reused by components
- Move JogController mixin to components, since it's used as a base
  class for components
- Implement StyledWidget as a base class for all components with custom
  style sheets, to avoid code repetition
- Remove mixins and helpers folders, since they are not needed anymore
