+++ src/desktop/desktop/application/connectionErrors.py
