+++ src/desktop/desktop/components/cards/FileCard.py
@@ -2,7 +2,7 @@ from core.domain.entities import File, Material, Tool
 from desktop.components.cards.Card import Card
 from desktop.components.dialogs.FileDataDialog import FileDataDialog
 from desktop.components.dialogs.TaskDataDialog import TaskFromFileDialog
-from desktop.helpers.utils import needs_confirmation
+from desktop.components.utils import needs_confirmation
 from PyQt5.QtCore import pyqtSignal
