+++ src/desktop/desktop/components/containers/WidgetsHList.py
