+++ src/desktop/desktop/components/gatewayMonitor.py
