+++ src/desktop/tests/components/containers/test_button_grid.py
@@ -1,19 +1,19 @@
-from desktop.containers.ButtonList import ButtonList
+from desktop.components.containers.ButtonGrid import ButtonGrid
 from PyQt5.QtWidgets import QPushButton
 from pytest_mock.plugin import MockerFixture
 from pytestqt.qtbot import QtBot
 
 
-class TestButtonList:
-    def test_button_list_init(self, qtbot: QtBot, mocker: MockerFixture, helpers):
+class TestButtonGrid:
+    def test_button_grid_init(self, qtbot: QtBot, mocker: MockerFixture, helpers):
         actions = [
             ("Action 1", mocker.Mock()),
             ("Action 2", mocker.Mock()),
             ("Action 3", mocker.Mock()),
             ("Action 4", mocker.Mock()),
         ]
-        button_list = ButtonList(actions)
-        qtbot.addWidget(button_list)
+        button_grid = ButtonGrid(actions)
+        qtbot.addWidget(button_grid)
 
         # Assertions
-        assert helpers.count_widgets(button_list.layout(), QPushButton) == 4
+        assert helpers.count_widgets(button_grid.layout(), QPushButton) == 4
